rename folder OhsomeQgis → ohsomeTools
