+++ ohsomeTools/gui/OhsomeQgisDialogUI.ui
@@ -181,7 +181,7 @@
            </property>
            <property name="icon">
             <iconset resource="resources.qrc">
-             <normaloff>:/plugins/OhsomeQgis/img/icon_refresh.png</normaloff>:/plugins/OhsomeQgis/img/icon_refresh.png</iconset>
+             <normaloff>:/plugins/ohsomeTools/img/icon_refresh.png</normaloff>:/plugins/ohsomeTools/img/icon_refresh.png</iconset>
            </property>
           </widget>
          </item>
@@ -201,7 +201,7 @@
            </property>
            <property name="icon">
             <iconset resource="resources.qrc">
-             <normaloff>:/plugins/OhsomeQgis/img/icon_settings.png</normaloff>:/plugins/OhsomeQgis/img/icon_settings.png</iconset>
+             <normaloff>:/plugins/ohsomeTools/img/icon_settings.png</normaloff>:/plugins/ohsomeTools/img/icon_settings.png</iconset>
            </property>
           </widget>
          </item>
@@ -380,7 +380,7 @@
                </property>
                <property name="icon">
                 <iconset resource="resources.qrc">
-                 <normaloff>:/plugins/OhsomeQgis/img/icon_add.png</normaloff>:/plugins/OhsomeQgis/img/icon_add.png</iconset>
+                 <normaloff>:/plugins/ohsomeTools/img/icon_add.png</normaloff>:/plugins/ohsomeTools/img/icon_add.png</iconset>
                </property>
               </widget>
              </item>
@@ -400,7 +400,7 @@
                </property>
                <property name="icon">
                 <iconset resource="resources.qrc">
-                 <normaloff>:/plugins/OhsomeQgis/img/icon_clear.png</normaloff>:/plugins/OhsomeQgis/img/icon_clear.png</iconset>
+                 <normaloff>:/plugins/ohsomeTools/img/icon_clear.png</normaloff>:/plugins/ohsomeTools/img/icon_clear.png</iconset>
                </property>
               </widget>
              </item>
@@ -504,7 +504,7 @@
                </property>
                <property name="icon">
                 <iconset resource="resources.qrc">
-                 <normaloff>:/plugins/OhsomeQgis/img/icon_clear.png</normaloff>:/plugins/OhsomeQgis/img/icon_clear.png</iconset>
+                 <normaloff>:/plugins/ohsomeTools/img/icon_clear.png</normaloff>:/plugins/ohsomeTools/img/icon_clear.png</iconset>
                </property>
               </widget>
              </item>
@@ -524,7 +524,7 @@
                </property>
                <property name="icon">
                 <iconset resource="resources.qrc">
-                 <normaloff>:/plugins/OhsomeQgis/img/icon_add.png</normaloff>:/plugins/OhsomeQgis/img/icon_add.png</iconset>
+                 <normaloff>:/plugins/ohsomeTools/img/icon_add.png</normaloff>:/plugins/ohsomeTools/img/icon_add.png</iconset>
                </property>
               </widget>
              </item>
@@ -661,7 +661,7 @@
                </property>
                <property name="icon">
                 <iconset resource="resources.qrc">
-                 <normaloff>:/plugins/OhsomeQgis/img/icon_clear.png</normaloff>:/plugins/OhsomeQgis/img/icon_clear.png</iconset>
+                 <normaloff>:/plugins/ohsomeTools/img/icon_clear.png</normaloff>:/plugins/ohsomeTools/img/icon_clear.png</iconset>
                </property>
               </widget>
              </item>
@@ -681,7 +681,7 @@
                </property>
                <property name="icon">
                 <iconset resource="resources.qrc">
-                 <normaloff>:/plugins/OhsomeQgis/img/icon_add.png</normaloff>:/plugins/OhsomeQgis/img/icon_add.png</iconset>
+                 <normaloff>:/plugins/ohsomeTools/img/icon_add.png</normaloff>:/plugins/ohsomeTools/img/icon_add.png</iconset>
                </property>
               </widget>
              </item>
@@ -1510,7 +1510,7 @@
         </property>
         <property name="icon">
          <iconset resource="resources.qrc">
-          <normaloff>:/plugins/OhsomeQgis/img/icon_help.png</normaloff>:/plugins/OhsomeQgis/img/icon_help.png</iconset>
+          <normaloff>:/plugins/ohsomeTools/img/icon_help.png</normaloff>:/plugins/ohsomeTools/img/icon_help.png</iconset>
         </property>
        </widget>
       </item>
@@ -1521,7 +1521,7 @@
         </property>
         <property name="icon">
          <iconset resource="resources.qrc">
-          <normaloff>:/plugins/OhsomeQgis/img/icon_about.png</normaloff>:/plugins/OhsomeQgis/img/icon_about.png</iconset>
+          <normaloff>:/plugins/ohsomeTools/img/icon_about.png</normaloff>:/plugins/ohsomeTools/img/icon_about.png</iconset>
         </property>
        </widget>
       </item>
