+++ ohsomeTools/gui/img/icon_clear.png
