+++ ohsomeTools/gui/img/icon_locate.png
