+++ ohsomeTools/gui/img/icon_ohsome.png
