+++ ohsomeTools/gui/img/icon_refresh.png
