+++ ohsomeTools/gui/resources.qrc
@@ -1,5 +1,5 @@
 <RCC>
-  <qresource prefix="/plugins/OhsomeQgis">
+  <qresource prefix="/plugins/ohsomeTools">
     <file>img/logo_heigit_300.png</file>
     <file>img/icon_add.png</file>
     <file>img/icon_refresh.png</file>
