+++ ohsomeTools/static/js/github_buttons.js
